% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dzik.R
\docType{data}
\name{Dzik}
\alias{Dzik}
\title{Wildboar genotypes.}
\format{
A genind object containing :
\describe{
  \item{82}{individuals}
  \item{12}{loci}
  \item{102}{alleles}
  \item{4}{populations}
}
}
\source{
private dataset
}
\usage{
Dzik
}
\description{
A dataset containing microsatellite data from wildboars.
}
\keyword{datasets}
