% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShannonGen-package.R
\docType{package}
\name{ShannonGen-package}
\alias{ShannonGen}
\alias{ShannonGen-package}
\title{ShannonGen: Estimation of the Shannon Diverstity Index and Average Weighted Nucleotide Diversity}
\description{
A package to calculate Shannon Diversity Index and Average Weighted Nucleotide Diversity.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/konopinski/Shannon/}
  \item Report bugs at \url{https://github.com/konopinski/Shannon/issues}
}

}
\keyword{internal}
