% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShannonGen.R
\name{ShannonGen}
\alias{ShannonGen}
\title{ShannonGen function \cr (modified code from \code{SpadeR})}
\usage{
ShannonGen(gInd, estimator = NULL)
}
\arguments{
\item{gInd}{a \emph{genind} object. For details on \emph{genind} objects see 
\href{https://github.com/thibautjombart/adegenet}{\code{adegenet}} documentation.}

\item{estimator}{a character string, a vector or a list of \emph{H} estimators that
will be estimated:
\itemize{
  \item\emph{ch}{ - Chao et al 2013,}
  \item\emph{cs}{ - Chao and Shen 2003,}
  \item\emph{sh}{ - the original Shannon diversity index (Shannon 1948),}
  \item\emph{z}{ - Zahl 1977 (\strong{default}).}
  }}
}
\value{
The function returns a list of data frames (one list element per each estimator)
  containing selected Shannon \emph{H} estimators for all loci. Rows represent loci, while columns
  represent populations.
}
\description{
ShannonGen function \cr (modified code from \code{SpadeR})
}
\details{
For the details on performance of Shannon \emph{H} estimators
 see Konopinski MK 2020 (\emph{in press})
 The functions \emph{MLE}, \emph{Z}, \emph{CS} and \emph{Ch} were copied from \code{Diversity} function
 from \href{https://github.com/AnneChao/SpadeR}{\code{SpadeR}} package
 (WARNING: the function was tested only on microsatellite data)
}
\examples{
data(Dzik) ## sample genotypes from wildboars

## Default usage:
ShannonGen(Dzik) ## returns Zahl 1977 unbiased estimator

## Calculating all the estimators:
ShannonGen(Dzik, estimator = c("z", "sh", "cs", "ch"))

}
\references{
Chao A., Ma K.H., Hsieh T.C. and Chiu C.H. 2016. SpadeR (Species-richness
     Prediction And Diversity Estimation in R): an R package in CRAN. Program and User's
     Guide also published at \href{https://github.com/AnneChao/SpadeR}{https://github.com/AnneChao/SpadeR}

Chao A. and Shen T.J. 2003. Nonparametric estimation of Shannon's index of
     diversity when there are unseen species. \emph{Environmental and Ecological Statistics}
     10:429-443. DOI: 10.1023/A:1026096204727

Chao A., Wang Y.T., Jost L. 2013. Entropy and thespecies accumulation curve:
     a nearly unbiased estimator ofentropy via discovery rates of new species.
     \emph{Methods in Ecology and Evolution} 4:1091-1100. DOI: 10.1111/2041-210X.12108

Konopinski MK. 2020.  Shannon diversity index: A call to replace the original
     Shannon's formula with the unbiased estimators. \emph{PeerJ, in press}

Shannon C.E. 1948. A Mathematical Theory of Communication.
     \emph{The Bell System Technical Journal}, 27:379-423,623-656.

Zahl S. 1977. Jackknifing An Index of Diversity. \emph{Ecology} 58 (4): 907-913.
     DOI: 10.2307/1936227
}
\keyword{ShannonGen}
