% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightedPi.R
\name{weightedPi}
\alias{weightedPi}
\title{weightedPi function \cr}
\usage{
weightedPi(sequences)
}
\arguments{
\item{sequences}{\emph{DNAbin} object.}
}
\value{
The function returns a value of average weighted nucleotide diversity for the sequences 
  for the samples included in \emph{DNAbin}  object.
}
\description{
This function estimates nucleotide diversity \emph{sensu} Nei and Li 1979 
  from \emph{DNAbin} objects. \emph{Fasta} files can be easily converted to \emph{DNAbin} using
  \code{ape::read.FASTA()} function with \code{type = "DNA"}. Check \code{ape} documentation for details.
  Should also work with fastq files but I didn't test that yet.
}
\references{
Konopinski MK. (unpublished) Weighted nucleotide diversity is more precise 
  than pixy in estimating true value of \eqn{\pi} from sequences containing missing data. 
  \emph{???} . DOI: \cr

Nei M and Li WH. 1979. Mathematical model for studying genetic variation 
  in terms of restriction endonucleases. \emph{PNAS}, 76(10), 5269–5273. 
  DOI: 10.1073/pnas.76.10.5269
}
\keyword{weightedPi}
